requirements.txt
# Apache Spark
pyspark==3.5.0

# Apache Airflow
apache-airflow==2.8.0
apache-airflow-providers-apache-spark==4.5.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==1.4.51

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Data Generation
faker==22.0.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
